+++ app/src/test/kotlin/me/xizzhu/android/joshua/core/ReadingProgressManagerTest.kt
@@ -24,7 +24,7 @@ import me.xizzhu.android.joshua.core.repository.BibleReadingRepository
 import me.xizzhu.android.joshua.core.repository.ReadingProgressRepository
 import me.xizzhu.android.joshua.tests.BaseUnitTest
 import me.xizzhu.android.joshua.tests.MockContents
-import me.xizzhu.android.joshua.utils.currentTimeMillis
+import me.xizzhu.android.joshua.utils.elapsedRealtime
 import org.mockito.Mock
 import org.mockito.Mockito.*
 import kotlin.test.BeforeTest
@@ -60,10 +60,10 @@ class ReadingProgressManagerTest : BaseUnitTest() {
         `when`(bibleReadingRepository.currentVerseIndex()).thenReturn(flowOf(VerseIndex(1, 2, 3)))
         `when`(bibleReadingRepository.currentTranslation()).thenReturn(flowOf(MockContents.kjvShortName))
 
-        currentTimeMillis = 1L
+        elapsedRealtime = 1L
         readingProgressManager.startTracking()
 
-        currentTimeMillis = ReadingProgressManager.TIME_SPENT_THRESHOLD_IN_MILLIS + 2L
+        elapsedRealtime = ReadingProgressManager.TIME_SPENT_THRESHOLD_IN_MILLIS + 2L
         readingProgressManager.stopTracking()
 
         verify(readingProgressRepository, times(1))