+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
   - Should call AndroidReadingProgressStorage.save() on IO thread
   - Do not fail if the requested verse does not exist when loading annotated verses
   - Do not fail when reading a large number of verses by verse indexes
+  - Should use elapsedRealTime() when tracking reading time
 - Updated dependencies:
   - Dagger to 2.25.4